refactoring: #7 패키지 분류 방법 변경
