+++ src/main/java/kr/kro/stamfood/restaurant/application/RestaurantRepository.java
@@ -1,0 +1,5 @@
+package kr.kro.stamfood.restaurant.application;
+
+
+public interface RestaurantRepository{
+}
