+++ src/main/java/kr/kro/stamfood/restaurant/application/RestaurantServiceImpl.java
@@ -1,0 +1,12 @@
+package kr.kro.stamfood.restaurant.application;
+
+import kr.kro.stamfood.restaurant.presentation.RestaurantService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class RestaurantServiceImpl implements RestaurantService {
+
+
+}
