+++ src/main/java/kr/kro/stamfood/restaurant/domain/RestaurantRepositoryImpl.java
@@ -1,0 +1,8 @@
+package kr.kro.stamfood.restaurant.domain;
+
+import kr.kro.stamfood.restaurant.application.RestaurantRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class RestaurantRepositoryImpl implements RestaurantRepository {
+}
