+++ src/main/java/kr/kro/stamfood/restaurant/presentation/RestaurantController.java
@@ -1,0 +1,4 @@
+package kr.kro.stamfood.restaurant.presentation;
+
+public class RestaurantController {
+}
