+++ src/main/java/kr/kro/stamfood/restaurant/presentation/RestaurantService.java
@@ -1,0 +1,3 @@
+package kr.kro.stamfood.restaurant.presentation;
+public interface RestaurantService {
+}
